app/requirements.txt
streamlit
streamlit-drawable-canvas
requests
Pillow
numpy

